+++ tui/init.go
@@ -1,0 +1,9 @@
+package tui
+
+import (
+	tea "github.com/charmbracelet/bubbletea"
+)
+
+func (m model) Init() tea.Cmd {
+	return m.spinner.Tick
+}
